+++ packages/lib/gpcircuits/tsconfig.json
@@ -9,7 +9,7 @@
     "typeRoots": ["./src/declarations"]
   },
   // Some artifact-including packages need to import a JSON file
-  "include": ["src", "test", "artifacts/*.json", "src/*.json"],
+  "include": ["src", "test", "artifacts/*.json", "src/*.json", "scripts/common.ts"],
   "exclude": ["dist", "build", "node_modules"],
   // DO NOT MODIFY MANUALLY BEYOND THIS POINT
   // References are automatically maintained by `yarn fix-references`
